interpreter: add chewing gum for stack traces of do-blocks
This restores the stack trace compromised in the previous commit, at
the cost of making a small mess. This is its own commit in order to be
able to revert it easily later.

This works by stuffing a Maybe String into VDo, populating it from the
stack trace goop, and feeding it to a wrapper around interpretStmts.
As an instance of chewing gum, it could be worse...