interpreter: move "force" out of interpretStmts.
Call it interpretMonadAction, and share it with the FromValue
instances so those don't need to paste copies of it any more.

(FUTURE: we could probably have one FromValue instance for all the
monads by using the InterpreterMonad class. But it's not completely
straightforward and the infrastructure may still change, so not yet.)