saw-script: improve the do-block representation in the AST.
Instead of just having a list of statements, the last of which is
required to actually be of the form _ <- e, separate out the final
expression. This allows simplifications in several places.

Invalid do-blocks that are empty or end with something that doesn't
provide a value are now rejected while consing in the parser, instead
of in the typechecker. The downside of that is that in the presence of
an invalid do-block one won't also get type errors, but that's not a
big deal.

The messages involved no longer contain the typechecker's fairly
useless "in x" addendum.

Fix a couple tests to not inadvertently contain invalid do-blocks.

This also revealed that the AutoMatch code was glued to the
interpreter incorrectly, such that if the last statement it generated
didn't happen to have the form "_ <- something", the interpreter would
crash executing it. Fix that. (Related to #2482.)