SAWScript: Add names to lambdas in the AST.
In order to get reliable, accurate, and complete stack traces, we need
the name of the function being called at the time the last argument is
applied. Because function declarations are always desugared into "let
f = lambda...", this means each lambda needs to know its name.

Do that. The name is a Maybe because there are also still anonymous
lambdas.

While here, and since I'm touching every use anyway, rename the
constructor in question from "Function" to "Lambda" for clarity.