interpreter: insert a layer of indirection around values for builtins
Add a BuiltinWrapper type that contains the closure for VBuiltin.

This doesn't do anything on its own, but it gives leverage on the
typeclass-generated code for wrapping builtins. In particular, we
want to be able to wrap a builtin in such a way that we can carry
around additional info in VBuiltin and retain it when applying
arguments, without making a mess.