interpreter: Construct builtin function values with BuiltinWrapper.
Completes what the previous commit started.

See the large comment attached to the IsValue class for a description
of how the new logic works and why it wasn't entirely trivial.

Add a VClosure value for values that need to hold closures where using
BuiltinValue isn't feasible. There are only a handful of these, and
they're all in the stack trace logic, which is scheduled for removal
in the very near future.

In the course of doing this, because every builtin that mucks directly
with function values needed attention anyway, I've rearranged the
logic for "subshell" and "proof_subshell" so as to fix #2306.