interpreter: Reform the runtime position handling.
It is not necessary for the global position to be valid when executing
in the interpreter, only when we leave the interpreter to execute in
builtins. Therefore, there's no need to push and pop it: we can just
assign it before each place we call out. The execution flow in the
interpreter is now more or less rational enough to be able to identify
these. (Although it still leaves a good deal to be desired.)

When leaving the interpreter to execute a Haskell-level monadic action
the position needs to be set in fromValue: if the action is an
argument to a builtin that's the last time the interpreter gets to do
anything. Note though that this means the position does not also need
to be set a second time when executing the action from inside the
interpreter. (While setting the position again is harmless, creating
extra stack trace frames won't be.)

It's also possible to re-enter the interpreter. There aren't any pure
(or "pure") callbacks, but a number of builtins take monadic values as
arguments and execute them. These are effectively callbacks, and
Haskell lazy evaluation means they don't run at all (including any
interpreted portion) until the builtin involved triggers that
execution. This is not a problem just for setting the position, but
will need to be accounted for when handling stack traces.

In particular there are implicit call sites inside builtins that we're
going to have to do something about. Not in this commit, though.

Within the interpreter none of this matters and can use positions
taken from the AST.

As part of these changes, carry call site positions into applyValue;
we'll need that for stack traces.

Meanwhile, the failure in test2049 is now reported on the correct line
number: the last call to ghost_value in the LLVM spec, instead of the
call to llvm_verify. Whee!