interpreter: One more piece of infrastructure.
We need Values that contain monadic actions to contain the position
where they were "called".

When we run fromValue on such a Value to get a Haskell-level monadic
action out, we might be in the interpreter, in which case the
interpreter can see to the position. But we could also be extracting
it to pass as an argument to a builtin. In that case, the builtin is
about to execute it (because of Haskell lazy evaluation, the fromValue
won't actually run until the monad action inside it is being used).
This is then the last (and perhaps only) chance we get to record that
we're executing in the monadic action.

In toValue instances, insert a placeholder position. This is what
will end up in the values in the builtins table.

Add a function injectPositionIntoMonadicValue to update the position
in any of these values. Call this to update the value position to the
current position in three critical places:

  - in interpretExpr, when looking in the environment;
  - in applyValue, on the argument value; and
  - when returning out the last expression in a do-block.

The first is required because in
   let solver = z3 in prove_print solver ...
we'll want call stack entries that point at the "z3" as well as the
"prove_print solver" program points.

The second is required so that if z3 took an argument and we had
   prove_print (z3 foo) ...
we'd get an entry for the "z3 foo"; we won't and shouldn't in that
case get one for just "z3" because it's not a VProofScript.

The third is required because the last expression of a do-block isn't
executed in the do-block and otherwise code of the form
   prove_print (do { ...; z3 }) ...
wouldn't get an entry for the "z3".

I think these are all the cases. If we ever end up seeing the
placeholder position in a stack trace, we might need to identify more.

This commit doesn't do anything yet, just carries the info around.

Note that we add a position to VReturn for consistency rather than
because it has any actual value: since return doesn't do anything,
nothing will ever look at the positions or stack trace entries it
logs.