interpreter: Add machinery for a new stack trace mechanism.
This commit provides the infrastructure for managing and updating (and
printing) the new stack traces, but does not include any of the code
for collecting the information. That's coming next.

Pass the current position when printing a stack trace, because that's
wanted for the new traces.

This commit prints an empty new stack trace along with every legacy
stack trace, so it needs to update every test reference output that
has a stack trace in it. The legacy traces haven't changed.