interpreter: Adjust handling of tails of do-blocks.
When executing a do-block, convert
   do { m1; m2; mr; };
to
   do { m1; m2; r <- mr; return r; }

that is, always execute the last action and return its result. This
makes no difference if it's already "return x", but if it does
something significant we execute it in the context of the do-block
it belongs to and not the parent context.

This in turn avoids a set of special cases for generating stack
traces, and, perhaps unsurprisingly, fixes some of the omissions in
the legacy stack traces found by the tests just added.