interpreter: Carry more metadata in monadic values.
For stack traces, we need more than the position. We need names as
well, and it appears we want a whole list of positions and names in
order to track things like "let x = print_stack; let y = x; y;" which
more or less are successive call sites.

VDo, however, doesn't actually need a position any more. Which I went
back and forth on repeatedly in the course of developing this
patchset. Sigh.

See enclosed comments for details.