interpreter: Have fromValue know its calling context.
fromValue can be called either from inside the interpreter or as a
result of unwrapping a function argument to pass to a builtin. The
difference is immaterial for ordinary values, but when passing a
callback (which in general is actually a plain monadic value) to a
builtin, it's necessary to know that this has happened. When
generating stack traces, we want an entry for the call to the
callback. Otherwise things can get quite confusing.

No functional change intended (yet).