interpreter: Add missing setPosition call.
There was supposed to be a setPosition call when applying the last
argument of a builtin, because we can leave the interpreter when that
happens.

(While a Haskell-level pure function can't call getPosition to examine
the result, "pure" SAWScript functions that actually execute in the
TopLevel monad might.)

This seems to have been accidentally left off the final version
of #2500.