manual: Work over the SAWScript chapter.
Or more accurately, write most of a new one. The old one is still here
after it.

XXX: this is not a finished commit; the old text needs to be merged in
XXX: or (probably mostly) removed.

XXX: also some of what's in here belongs in the SAWScript reference
XXX: chapter, except it doesn't exist yet.