manual: Work over the SAWScript chapter.
Or more accurately, write most of a new one. The old one is still here
after it.

That's going to need to be cleaned up later. (And, speaking from
later: I'd been intending to retcon this, but there have been too many
additions/adjustments to this chapter in the meantime. So this
half-finished version is going to remain in the history for posterity.)