tests: Track provenance for binary blobs with separate *.FROM files.
Provide scripts check-from.sh and update-from.sh for, respectively,
validating these files and updating the version info in them.

The motivation for moving to one tracking file per blob is that we
want to be able to automatically generate this info when we rebuild
those blobs, and that's much more easily done one file at a time.

(We have been storing the info as comments in the Makefiles, which is
much harder to process automatically.)

The source of the information in these files is:

 - For the linked-mir.json files that got regenerated in #2550 and
   didn't already have provenance info, the known provenance info
   from that rebuild.

 - For the LLVM bitcode and Java class files in the Python tests,
   which date back a ways, placeholder files saying that the
   provenance is unknown.

 - For the test_llvm_x86_* LLVM bitcode files that were added a
   few commits back, the known provenance info from that build.

 - For the other preexisting blobs in test_llvm_x86_*, placeholder
   files saying that the provenance is unknown.

 - For the test_crucible_jvm Java class files that were added a few
   commits back, the known provenance info from that build.

 - For the other preexisting blobs in test_crucible_jvm, placeholder
   files saying that the provenance is unknown.

 - For these tests, it was moved from comments in the Makefile:
      test0220 test1977 test1983 test1998_mir test2032 test2049
      test2122_typecheck_gap test2148 test2257 test2363 test2398
      test_llvm_errors test_mir_unsound_global test_stack_traces