tests: Add missing .gitignore files.
 - test0002
 - test2122_typecheck_gap
 - test2314
 - test_unint_extcore

Also remove some unwanted upper-level gitignores. It's perhaps tidier
to have one .gitignore file in intTests, but it hasn't been maintained
and it's easier to keep track of test-specific bits like this in each
test dir.

Note in passing that for test0002, the output files are deleted on
success; however, they may be left behind on failure, so they should
still be ignored.