tests: Add some missing .gitignore files.
(Remove one unused .gitignore file, too.)

Also move per-test ignore logic to the test directories. It's perhaps
tidier to have one .gitignore file in intTests that covers everything,
but it hasn't been maintained and it's easier to keep track of
test-specific bits in each test dir.

Note: For test0002, the output files are deleted on success; however,
they may be left behind on failure, so they should still be ignored.