tests: Check in LLVM bitcode files for test_llvm_x86_*.
Add makefiles in test_llvm_x86_* to regen them (and the other bits)
per current practices.

This moves the artefact build bits out of test.sh. Once upon a time
they were all built on the fly. That then didn't work on non-x86
targets, at which point some of the artefacts (but not all) got
checked in and their build actions commented out in test.sh.

However, the bitcode files were still being generated on the fly.
This used whatever clang you happened to have, for whatever target it
happened to default to, and was thus only working on non-x86 targets
by accident.

The provenance for these bitcode files has not been forgotten and will
be added in an upcoming commit.