interpreter: Don't run main without checking its type.
It seems inadvisable to require any value we find called "main" to
have type TopLevel (), because the "main" feature isn't generally or
widely used and many users probably aren't aware of it.

So, like if we do nothing if there's no main, do nothing unless the
type for the main we find looks like something monadic. In that case,
check it against {a} TopLevel a and complain if it doesn't match.
Unfortunately we seem to need to accept things like TopLevel Theorem
and can't just require TopLevel ().

Doing this reliably requires a new typechecker entry point,
unfortunately, because the type of main might quite plausibly be
polymorphic and we want to use the unification logic to cope with it.

Add some tests.

Closes #2556.