tests: Add makefiles for tests that leave files behind.
Every test with a .gitignore to ignore things produced by running said
test now also has a Makefile with a clean rule to remove those things.

It is possible that some tests that leave things behind don't have
either Makefiles or .gitignore files, but we can fix them when and if
they surface.