Create the branch point for SAW 1.4.
This updates the version number to 1.3.0.99.99 (which is a little
silly, but it's not 1.4 yet and there will be more changes that get
merged into the release branch before the release happens) and adds
the 1.4 heading in CHANGES.md. Note that it does _not_ yet set the
release date because that's still "soon".