crucible-mir-comp: Alloc correct size in poststate
Previously, if there was a fresh allocation in the post state of a
MethodSpec, the override allocated space for a single value of the
MirAllocSpec's type and ignored the MirAllocSpec's length. Fix this by
always allocating a MirVector of the correct length, whose element type
is the MirAllocSpec's type, the same way it is done in SAW.

This fixes one of the bugs identified in #2477.