REPL: Add the ability to tab-complete SAWScript types.
Only works for :search, because we'd need to reimplement the parser to
be able to know when we're in a context expecting a type. As it stands,
we only have a reimplementation of enough of the lexer to recognize
string literals and {{ }} {| |} blocks.

(And the completion logic assumes all string literals are filenames.)