MIR backend: Properly support const generics
This contains the SAW-side changes needed to properly support const generics in
the MIR backend, as detailed in
https://github.com/GaloisInc/mir-json/issues/64. Specifically, this adds a
`mir_const : MIRType -> Term -> MIRType` function that can be used to specify a
constant used to instantiate a const generic parameter. The `MIRType` that this
returns can then be passed to `mir_find_adt` to look up an instantiation of a
const generic ADT. (See the new text in the SAW user manual for limitations of
this approach.)

This bumps the `mir-json` submodule to bring in the changes from
GaloisInc/mir-json#173, as well as the `crucible` submodule to bring in the
changes from https://github.com/GaloisInc/crucible/pull/1552. Because this
requires a bump to the MIR JSON schema version, corresponding changes to the
version number must also be made here.