traverseMirAggregate: Add a special case for empty MirAggregates
Doing so is necessary to work around a bug in the implementation of
`IntMap.mergeWithKey` that exists in pre-0.8 versions of the `containers`
library (https://github.com/haskell/containers/issues/979), which all currently
supported versions of GHC currently use. See the newly added comments in
`traverseMirAggregate` for the full story.

Fixes #2687.