typechecker: Move the type environment from RO to RW.
This commit has just this move and none of the ensuing cleanup it's
supposed to enable. Trying to do all of it at once broke something,
so I'm subdividing into smaller steps.

(Having the environments in the RO portion of the monad implies that
nothing ever updates either environment and that the language only has
scoped let-bindings. This is a bald-faced lie and the existing code
works around the consequences by throwing around continuations like
they're candy.)