typechecker: Fix up the environment manipulation infrastructure.
In particular we want both "with..." and "add..." versions that wrap
and don't wrap (respectively) a sub-action, because the language has
both nesting and updating environment constructs.

In the long run it may be better to teach the environment structure
itself to know about scopes instead of saving and restoring the
environment global variable in the monad for each nested construct;
things are usually cleaner that way. But not yet.

Note: the new entry points to this stuff are unused for the moment
because I'm doing the change to deploy it separately in the next
commit.