support: Add a ScopedMap type.
This is like Data.Map except it's a chain of maps that function as
scopes; keys in inner (more recent) scopes override matching keys in
outer (older) scopes, as one wants for scoped variable lookup.

You can also push and pop the scope entries as you carry the map in
and out of scoping constructs.

Also add SAWSupport.Panic because there isn't one yet and ScopedMap
has a panic case: trying to pop the last scope.