Simplify the ecdsa example a bit to avoid stepping on #2726.
This removes uses of "let {{ thm = ... }}" followed by explicitly
binding and unfolding "thm" with just binding a Cryptol-level lambda,
like other code in the same file. This is simpler and probably better
for an example regardless of other considerations.

But also, the "let {{ thm = ... }}" binds a SAWCore-level name "thm",
which causes errors about redefining it because the SAWCore
environment doesn't get SAWScript-level scopes. That's #2726.

After the upcoming fixes so the Cryptol environment _does_ get
SAWScript-level scope handling it does, anyway. It isn't obvious why
this code didn't already exhibit that behavior. Perhaps there's
something down inside that overwrites the corresponding SAWCore
definition when you overwrite a Cryptol definition, which no longer
applies since we no longer do that. (And maybe that logic checks to
make sure the SAWCore manipulation is sound, and maybe it doesn't...)

Regardless, we have at least one other test that explicitly
captures #2726.