interpreter: Add CryptolEnvStack.
A CryptolEnvStack is a stack of CryptolEnv; the idea is that the stack
allows pushing and popping the Cryptol environment at runtime to match
SAWScript scopes.

This adds the possibility of having the Cryptol environment track the
SAWScript scopes as they come and go, in pursuit of #2304. However,
it's not hooked up yet. The interpreter environments cleanup needs to
happen before that's feasible; however, making these changes first
allows them to appear as their own commit.

(It was found that fixing the interpreter's environment handling
demanded also fixing the Cryptol environment scoping. Otherwise a lot
of things break. The interpreter's old environment handling defers
updating the Cryptol environment in some circumstances, which is
apparently enough to get a closer approximation of correct behavior in
the Cryptol environment than a rational approach. Because of this,
these changes were first written on top of the environment handling
cleanup and then moved before it. Because of _that_, this commit has
not been tested with a full CI run, because GH won't run actions on
commits that aren't the tips of pull-request branches. I've taken some
precautions, and I _think_ I've kept every access to the Cryptol
environment the same with respect to using "getMergedEnv", but I might
have fumbled one, or something else might break. Anyway, there's some
risk of this commit turning out to be non-testable in the future. It
is not _supposed_ to make any functional change.)

Note that there are a couple new error paths where attempting
import-like operations in a nested scope will fail; these are not
reachable until the scopes are connected up.

Besides that:

 - Add more accessors for the Cryptol environment and deploy them to
simplify the accesses in the REPL and in saw-server. This avoids
baking in exactly where and how the Cryptol environment appears in
TopLevelRW (note that saw-server shouldn't be using TopLevelRW at all
anyway...)

 - Canonicalize some of the related naming:
      - The canonical variable name for a CryptolEnv will be "ce" or
        "cenv", plural "ces" or "cenvs".
      - The canonical name for a CryptolEnvStack will be "cryenvs".

 - Tidy some of the adjacent code.