Normalize imports of Data.Map, Data.Set, etc.
These should be done like this:
   import qualified Data.Map as Map
   import Data.Map (Map)

and since nearly all but not quite the whole codebase already does
this, the exceptions are irritating and lead to wasted time.

(This has been bugging me for a while, relatively minor though it is.)

Add a place to document this, and in the future other such conventions.

No functional change intended.