saw-core: Skip extra WHNF step in SCTypeCheck.applyPiTyped.
Such reductions can cause types on Pi binders to be rewritten,
so that they no longer match the types annotated on occurrences
of the bound variable.

This in turn can cause type checking to fail, as one of the
enforced invariants is that all 'Variable' subterms in the scope
of a lambda or Pi should have a type that matches exactly.