REPL / Interpreter: Rationalize the way subshells are started.
The functions to start a subshell have to be passed down into the
interpreter (otherwise there's a nice circular dependency from the
interpreter back into the REPL) ... but they can be done so in a
much more reasonable way than had previously been happening.

As a side effect, the message "Proof subshells not supported" has been
moved to the dustbin of history. You can now always start either kind
of subshell and you don't need special permission. Mention this in
the changelog.