intTests: Extend test-and-diff.sh to handle providing stdin.
Switch test2242_2243 to use test-and-diff.sh.

While working on the previous changes I became concerned about whether
failures would actually be detected; the answer proves to be no.
Errors that occur in a nested interactive subshell get trapped and
discarded (because that's what the REPL does with errors) and at least
present01 and present03 do not fail overall if the actions in the
nested subshell, which are the point of those tests, fail.

This requires adding some mangy stuff to the output postprocessing so
that the timestamps in the output actually go away; however, we can
get rid of all that once we fix #2090.