SAWScript.Import: Return in Maybe instead of throwing.
For now, shift the throws to the callers. This might get rearranged
later, but for the moment I'm about to add more stuff to this module
that has different caller requirements and I'd like it to all be
uniform.

The state of the chewing gum currently serving as SAW's exception
handling is such that this changes the way parse errors print. (The
whitespace is different and there's now a vacuous stack trace like
type errors generate.) Accept this for now and update all the affected
test reference outputs.