saw-central: Avoid unfolding primitives with `goal_normalize`.
Some SAWCore constants (e.g. `equalNat`) are implemented as primitives
in the SAW backends, and also have definitions using e.g. Nat#rec that
are used for Coq export.

We now avoid unfolding such constants with `goal_normalize` because
it typically results in goals that are unprovable by external solvers.