SAWScript: Move everything that loads SAWScript text to Import.hs.
Now it's the only place that calls directly into the lexer and parser.

There wasn't any code that was doing anything inappropriate with the
lexer output, or anything besides feeding it directly to the parser;
but on the other hand there's no reason to have five copies of that
logic across three modules.

NFCI