SAWScript.Import: Typecheck after parsing type schemas.
There was no check in the caller; we were assuming all of the types of
builtins were valid, and as noted in the previous commit one of them
wasn't. "Oops"

Beware: the check is only evaluated if something actually looks at the
type. Running :search or :env is sufficient (and we have tests that do
that) but just starting SAW with a bad type in the builtins table will
not panic.

(I tried adding some strictness notations and that didn't help; but I
think detecting problems on any :search or :env run is probably
acceptable so didn't try that hard.)