SAWScript.Import: Flip the sense of the naming.
Instead of calling out the functions that call the typechecker as
"checked", call out the ones that don't as "unchecked". NFCI.

Currently the only uses that do not call the typechecker are the ones
that lead into the interpreter's statement-at-a-time execution, and
that's going to require other more involved cleanup before it can be
made to go away.