Support up to LLVM 20
This bumps the `llvm-pretty`, `llvm-pretty-bc-parser`, and `crucible`
submodules in order to bring in several downstream changes that, in total,
allow SAW to support LLVM versions up to 20. These include:

* `llvm-pretty`:
  https://github.com/GaloisInc/llvm-pretty/pull/168,
  https://github.com/GaloisInc/llvm-pretty/pull/169, and
  https://github.com/GaloisInc/llvm-pretty/pull/170

* `llvm-pretty-bc-parser`:
  https://github.com/GaloisInc/llvm-pretty-bc-parser/pull/316,
  https://github.com/GaloisInc/llvm-pretty-bc-parser/pull/317, and
  https://github.com/GaloisInc/llvm-pretty-bc-parser/pull/318

* `crucible`:
  https://github.com/GaloisInc/crucible/pull/1600,
  https://github.com/GaloisInc/crucible/pull/1602,
  https://github.com/GaloisInc/crucible/pull/1603, and
  https://github.com/GaloisInc/crucible/pull/1606

The only code changes that had to be made in SAW itself involve the LLVM
backend's skeleton-related commands, which look up debug metadata to figure out
which variables are declared. I have updated this code to also look at debug
records when using newer versions of LLVM.