saw-core: scWhnf doesn't do extra reductions on type expressions.
This makes it more like the usual definition of WHNF.

This also ensures that scWhnf does not break Term invariants about
type annotations on bound variables. The old definition of scWhnf
would modify the type of the input variable, but it was not guaranteed
to update the types on occurrences of that variable in the body of
the term.