SAWScript: Remove LLVMSetup/JavaSetup/MIRSetup hardwired types.
It turned out that while the hardwired type for Java verification
setup was "JavaSetup", we've actually been using the type "JVMSetup",
which is an ordinary abstract type and not a magic hardwired one, for
all the Java verification builtins and all actual proofs.

It is obviously wrong to have an additional hardwired JavaSetup type,
so remove it. This closes #2768.

The JVMSetup type, however, serves as proof that the magic hardwired
types aren't needed, so remove the LLVMSetup and MIRSetup ones as well
in favor of ordinary abstract types in the builtin types table.

"LLVMSetup", "JavaSetup", and "MIRSetup" are no longer reserved words.

The old name "CrucibleSetup" for "LLVMSetup" is, for the time being,
still a reserved word; it could be a typedef but there are certain
complications with having builtin typedefs that I don't want to get
into right now, namely the need to have the expansions appear in the
typing environment for checking. This should be tidied up at some
point in the future.