SAWScript: Set up to allow abstract types of kind beyond '*'.
This allows proper handling of the abstract types that are or are now
monads (LLVMSetup, JVMSetup, MIRSetup) but doesn't actually do it yet
because kind checking is bust. It will be wanted with the changes to
fix kind checking, though.

Move the kind representation to the AST (it was previously hidden in
the typechecker) so it can be stashed in the AbstractType case of
NamedType.