Typechecker: have generalize update patterns as well as exprs.
Otherwise the types left behind in the patterns are trash.

Not related to the current problems, but seems like a good thing to
fix.

XXX: They still seem to be wrong; they're no longer loose unification
XXX: vars, but they're also type variables that aren't the ones
XXX: forall-bound in the type scheme. I suspect one of the various
XXX: issues with the substitution mechanism already cited in comments.