CI: Bump MacOS versions. Move to what4-solvers-20251112.
The intel build is bumped from MacOS 13.x to 15.x.
The arm64 build is bumped from MacOS 15.x to 15.x.

There are supposedly macos-14-intel runners, but they do not appear to
actually exist; an attempt to use them sat for 24 hours before timing
out, without ever actually being scheduled.

There are still arm64 MacOS runners, but it seems weird for the intel
build to be on a newer version than the arm64 version.