interpreter: Reorganize builtins table.
I've tried to group everything by category and pick a reasonable
semantic ordering both of categories and within categories. Probably
not the last word, and there's certainly lots of room to kibitz, but
it should be a noticeable improvement on the status quo.

No real changes, only moving chunks of text around.