saw-core-coq-tests: golden-style prelude test.
Replaces the previous prelude test with a golden-style test that checks the
translated SAWCore prelude against a known-good file using the
`intTests/support/test-and-diff.sh` script. The previous test effectively only
tested that translating the prelude didn't cause a crash, so this test should
cover a superset of that functionality.