SAWScript parser: Tidy the entry points.
- G/C unused entry point for parsing a single statement
- Fix the REPL's entry point to accept multiple statements
- Share logic with the whole-file entry point
- Adjust names accordingly

Since we just changed the downstream plumbing to accept multiple
statements, this costs nothing. (The REPL continues to not demand a
final semicolon; the whole-file parser still does.)

This eliminates both of the following annoying traditional behaviors:

   sawscript> // hi
   Parse error at <stdin>:2:1-2:4: Unexpected `EOF'
   sawscript> print 3; print 4
   Parse error at <stdin>:1:10-1:15: Unexpected `print'

Instead it works:

   sawscript> // hi
   sawscript> print 3; print 4
   [02:24:47.040] 3
   [02:24:47.041] 4

Remove some workarounds for the comment-handling bug in the search
tests.