SAWScript: Process includes immediately after loading.
This allows all the program text to be visible to the typechecker,
and eliminates the assorted problems described in #2760.

We still allow includes in nested scopes. There's no clear reason
to allow that, but also no clear reason to prohibit it either.

Note that we need to load include names that are relative paths
relative to the directory of the file containing the include. We
can do this relatively easily now.

However, because #1791 is not just about includes but also other file
accesses, we still need the interpreter to chdir around, so we need to
tell it where to go. Otherwise the information is lost. Wrap each
include file in a pushdir/popdir pair to accomplish this. Note that
the interpreter also still wraps each file execution in a try/catch
that saves and restores the current directory and the directory stack.
This will prevent us from going off the rails if something throws
while we're executing an included file.

This whole scheme is kind of unsatisfactory (it's at least untidy) but
there's going to be a lot more work needed before we can deal properly
with #1791.

Remove the existing chdir operation in the interpreter, because it
doesn't accomplish anything any more.

Note that another cute side effect of these changes is that now, if
you ask the :type REPL command for the type of a do-block that has
an include in it, it'll work instead of ignoring the include.