MIR backend: Avoid panicking when constructing override-related error messages when matching enums
The previous error messages would attempt to pretty-print the entire enum value
when it hadn't been validated yet, and printing the enum could panic under
certain situation. There's not really a good reason to print the entire enum
here, however, as we can just print the specific parts of the enum that are
being checked (namely, the discriminant and the variant validity predicate).
Doing so avoids panicking.

Fixes #2799.