SAWScript: Add a pushdir/popdir and directory stack implementation.
We're about to change things so includes get resolved prior to
execution; but as things stand we need to be able to chdir to the
location of each included subscript as we execute it. The least
destructive way to arrange that seems to be to insert pushdir and
popdir operations around each included file's contents when we load
same.

This logic is internal-only; don't expose user-facing pushdir and
popdir implementations. The interaction of their behavior with
"include" (not to mention with exceptions) would be complicated to
explain and probably tricky to use.

Nothing uses this yet. Coming right up...