typechecker: Switch to ScopedMap for the naming environments.
This is the first step only that switches the types out; we aren't
using the ScopedMap scopes yet. However, it should all still work
this way and therefore this should be a separate commit.

Note that when the interpreter munges up its environments to pass to
the typechecker, we end up flattening the interpreter's ScopedMap so
that the separate environment for rebindable variables can be merged
into it. It would be possible to adjust the typechecker to also use a
separate environment for rebindables. However, after consideration it
seems like this is not desirable; the rebindable environment is
separate in the interpreter to make updating it work correctly (it
cannot be part of what gets closed in with lambdas) but this doesn't
apply to the typechecker.

All that munging logic will go away in the near term: we don't want
the interpreter and typechecker sharing environments at all.