yosys_import: Support JSON produced with -compat-int
To do so, we parse cell parameter values more generally to accept all possible
JSON values, and then we refine the code which parses `*_SIGNED` parameters to
accept both strings (which is what Yosys's `write_json` command normally
produces for parameters) and numbers (which is what is produced when using
`write_json -compat-int`).

Fixes #2846.