saw-core: Split "trusted kernel" of SharedTerm into Term.Certified.
The SharedContext type, all basic term former operations, and
all basic SharedContext operations are now defined in module
SAWCore.Term.Certified, which exports types SharedContext and
Term as abstract datatypes. Module Term.Certified is now
considered to be the "trusted kernel" of SAWCore.

Module SAWCore.SharedTerm is now layered on top of Term.Certified,
and defines various derived term operations that do not need to
be defined inside the kernel.