Fix goal_(intro|retract)_hyp sanity check
Previously the normalization rules could never fire for and, or, not, implies because expressions were reduced using scWHNF and all of the above functions a prelude definitions that evaluate away to underlying primitives.

Additionally the sanity check had the direction flipped when checking that the proven goals were subsumed by the needed goals. This was probably never noticed because the two lists were typically equal.