saw-core: Optimize conversions for performance and correctness.
Function `globalConv` uses the `Conversionable` type class
to generate a conversion from a Haskell function based on
its type. With the new binary representation for numeric
literals, the matcher for Haskell type `Int` is now relatively
expensive, so we improve performance by replacing type `Int`
with `()` for any unused arguments, so that the generated
conversions won't waste effort inspecting the SAWCore terms
in those positions.

We also make conversions for `at` conditional on the index
being in bounds. Previously the `at_VecLit` conversion would
raise a runtime error if it encountered an out-of-bounds index,
and `at_bvNat` would incorrectly rewrite out-of-bounds lookups
to `False`. Now, both of those conversions will refuse to
rewrite `at n a v i` if the index `i` is out of bounds.