cryptol-saw-core: Typecheck the innermost expression in a list comprehension with a known type
When typechecking `[e | x <- xs] : [n]ty`, we were inferring the type of `e`,
which means that if `e`'s type wasn't the same as `ty`, then SAWCore would
reject it when building up the type of the overall expression. Inferring `e`'s
type is silly, however, as we know that its type should be `ty` in the end. As
such, we now check `e : ty` using `importExpr'` (instead of inferring it with
`importExpr`), which ensures that we cast `e` to type `ty` with `coerce` if
necessary.

Fixes #2865.